V1.0.5 release
V1.0.5 release now supports a custom Image folder in the Application Support directory and makes it possible for end users to add modified images. When the folder is not found it will use the default images from the app bundle, when the folder is found it expects the images in the Application Support folder, of course one can rip the images from the RestaurantServer.